everything working now. Including indirect columns.
